Add an OS check to leakcheck()

+++ .bash_profile
@@ -30,10 +30,12 @@ function kill_emacs() {
 
 # Temporary alternative to Valgrind
 function leakcheck() {
-	# See advice about symlink given here: https://www.reddit.com/r/macprogramming/comments/7iao6d/using_the_leaks_command_on_a_cc_executable/
-	sudo ln -s /Applications/Xcode.app/Contents/Developer/usr/lib/libLeaksAtExit.dylib /usr/local/lib/libLeaksAtExit.dylib
-	leaks --atExit -- "$1"
-	rm /usr/local/lib/libLeaksAtExit.dylib
+	if [ "$(uname)" = "Darwin" ]; then
+		# See advice about symlink given here: https://www.reddit.com/r/macprogramming/comments/7iao6d/using_the_leaks_command_on_a_cc_executable/
+		sudo ln -s /Applications/Xcode.app/Contents/Developer/usr/lib/libLeaksAtExit.dylib /usr/local/lib/libLeaksAtExit.dylib
+		leaks --atExit -- "$1"
+		rm /usr/local/lib/libLeaksAtExit.dylib
+	fi
 }
 
 # Prompt
